feat: Move Baggage from public API to internal API
